remove .float()

+++ torchtitan/models/llama3/model/model.py
@@ -421,5 +421,5 @@ class Transformer(nn.Module, ModelProtocol):
             h = layer(h, self.freqs_cis)
 
         h = self.norm(h) if self.norm else h
-        output = self.output(h).float() if self.output else h
+        output = self.output(h) if self.output else h
         return output
